pyproject.toml
[tool.poetry]
name = "account-stats"
version = "0.1.0"
description = ""
authors = ["Gregorie <grishagti@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
gspread = "^6.1.2"
google-auth = "^2.34.0"
pydantic = "^2.9.2"
psycopg2-binary = "^2.9.9"
openpyxl = "^3.1.5"
python-telegram-bot = "13.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
